restructured AI Planner files
