+++ dotnet/packages/Microsoft.Bot.Builder.M365/Microsoft.Bot.Builder.M365/AI/Planner/ParsedCommandResult.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Microsoft.Bot.Builder.M365.AI
+namespace Microsoft.Bot.Builder.M365.AI.Planner
 {
     public class ParsedCommandResult
     {
