+++ dotnet/packages/Microsoft.Bot.Builder.M365/Microsoft.Bot.Builder.M365/AI/Planner/PredictedSayCommand.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Microsoft.Bot.Builder.M365.AI
+namespace Microsoft.Bot.Builder.M365.AI.Planner
 {
     public class PredictedSayCommand : IPredictedCommand
     {
@@ -17,5 +17,5 @@ namespace Microsoft.Bot.Builder.M365.AI
             Response = response;
         }
     }
-    
+
 }
